feature
getting the tennis environment running with self play and past all the errors having to do with numpy. The agents aren't really learning though. Originally the random seed was at 5 and changing it to 0 actually had a significant impact on the intial paddle movements between the two runs I did, not sure how much weight to put on that as a strategy. However, while I watched several episodes on eval mode and saw the paddles change their behavior over time the most reward they got in an episode was basically the equivalent of getting the ball over the net one time. So need to look at why they aren't improving past that point.